+++ src/shinestacker/algorithms/utils.py
@@ -9,31 +9,30 @@ import numpy as np
 import cv2
 import matplotlib.pyplot as plt
 from .. config.config import config
-from .. core.exceptions import ShapeError, BitDepthError, PathTooLong
+from .. core.exceptions import ShapeError, BitDepthError, PathTooLong, InvalidWinPath
 
 
 def get_path_extension(path):
     return os.path.splitext(path)[1].lstrip('.')
 
 
-def check_path_length_win(path):
+def check_windows_path(path):
     if not sys.platform.startswith('win'):
-        return True
-    abs_path = os.path.abspath(path)
-    long_paths_enabled = False
+        return
     try:
-        import winreg
-        with winreg.OpenKey(winreg.HKEY_LOCAL_MACHINE, 
-                            r"SYSTEM\CurrentControlSet\Control\FileSystem") as key:
-            print("key: ",  winreg.QueryValueEx(key, "LongPathsEnabled")[0])
-            long_paths_enabled = winreg.QueryValueEx(key, "LongPathsEnabled")[0] == 1
-    except Exception as e:
-        traceback.print_tb(e.__traceback__)
-    if long_paths_enabled:
-        return True
-    if len(abs_path) >= 260:
-        return False
-    return True
+        path.encode('ascii')
+    except UnicodeEncodeError:
+        raise InvalidWinPath(path)
+    abs_path = os.path.abspath(path)
+    if len(abs_path) > 260:
+        try:
+            import winreg
+            with winreg.OpenKey(winreg.HKEY_LOCAL_MACHINE,
+                                r"SYSTEM\CurrentControlSet\Control\FileSystem") as key:
+                if winreg.QueryValueEx(key, "LongPathsEnabled")[0] == 0:
+                    raise PathTooLong(abs_path)
+        except Exception:
+            raise PathTooLong(abs_path)
 
 
 EXTENSIONS_TIF = ['tif', 'tiff']
@@ -92,8 +91,7 @@ def extension_supported(path):
 
 
 def read_img(file_path):
-    if not check_path_length_win(file_path):
-        raise PathTooLong(file_path)
+    check_windows_path(file_path)
     if not os.path.isfile(file_path):
         raise RuntimeError("File does not exist: " + file_path)
     img = None
@@ -105,8 +103,7 @@ def read_img(file_path):
 
 
 def write_img(file_path, img):
-    if not check_path_length_win(file_path):
-        raise PathTooLong(file_path)
+    check_windows_path(file_path)
     if extension_jpg(file_path):
         cv2.imwrite(file_path, img, [int(cv2.IMWRITE_JPEG_QUALITY), 100])
     elif extension_tif(file_path):
@@ -137,7 +134,7 @@ def img_bw(img):
 
 def get_first_image_file(filenames):
     if len(filenames) == 0:
-        raise ValueError("No image files found in the selected path")
+        raise ValueError("No valid image files found in the selected path")
     first_img_file = None
     for filename in filenames:
         if os.path.isfile(filename) and extension_supported(filename):